Adds confluet-kafka-go as goduck.Stream implemention.
This implementation fixes some consistency problems related to rebalancing
and slow consumers.

This library depends on librdkafka, which comes statically bundled with
confluent-kafka-go. Since it depends on CGO, the only way it is possible to
build statically linked binaries is using musl-libc in alpine.

See confluent-kafka-go repository for more information on building.

Links:

 - https://github.com/confluentinc/confluent-kafka-go