Exposes metadata to the Processor
Messaging systems typically allow messages to have arbitrary metadata attached
to them. But with the current goduck implementation, it's not possible to
access those values, since it only exposes a []byte value.

A typical usecase would be to attach trace_id to the message headers, instead
of the body. Also, a correct kafka DLQ middleware is only possible when the
DLQ message is produced with the original messages's key.

In order to achieve this, a new type "goduck.Message" was created, that
encapsulates Value + Metadata. The message's key (specific for kafka) is a
special metadata accessible by msg.Metadata[goduck.KeyMetadata]

This is a breaking change, since the Processor contract was changed. Some
refactoring is required upon upgrading.

Rejected alternative:
A more elegant solution would be to use context variables to store that
information. But this solution fails with BatchStreamEngine, where the same
BatchProcessor handles multiple messages.