Refactors impl/ folder
The impl/ folder was using an old pattern, where every implementation would
fall into the same package. The major problem with this approach is that it
forces the clients to import way more dependencies than it is really needed.

This commit refactors the implementations into specific folders, while also
mantaining the deprecated versions for compatibility.

Fixes #11