Adds support to multiple elasticsink messages (#22)
The current SinkMessage type is only good for Indexing. For other
operations we need more types of messages

This commit deprecates SinkMessage in favor of IndexMessage and
DeleteMessage. Now the encode function must create the proper kind of
message.

The SinkMessage is now an alias to IndexMessage to maintain
compatibility, but it will be removed in future a version.