Adds a New stream func that does not panic (#19)
Previously existed only the MustNew function for creating streams. This
function panics in case of errors. This can be a problem sometimes.

This commit adds a new function New that returns an error instead of
panic and refactors MustNew to call New and panic in case of an error.

Also fix the error when topic was empty. It was reporting empty broker
instead of empty topic.

Ajust error messages to use 'bad' instead of 'invalid' to conform with standard.