Adds bigquery sink for pipeline
This is a pipeline.Sink that will save the messages to the Bigquery.

The sink can be configured to be more verbose by adjusting the log level
and also can ignore some errors, like invalid messages.

A couple of tests were broken and were fixed.

All dependencies have also been updated.