Adds pubsub support to pipeline
Makes the pipeline works with message pool and adds a new configuration
in the Config. It automatically creates a message pool and starts a job
pool engine based on the configuration.

WithConfig option was updated to configure a message pool if the
configuration allows it.

The build function will start a job pool engine if a message pool is configured.