Adds pubsub support to pipeline (#43)
Makes the pipeline works with message pool and adds a new configuration
in the Config. It automatically creates a message pool and starts a job
pool engine based on the configuration.

WithConfig option was updated to configure a message pool if the
configuration allows it.

The build function will start a job pool engine if a message pool is configured.

This commit also format files to conform with the new linting rules.