Adds new interfaces for handling messages
The current goduck model isn't very extensible, since it only handles message content ([]byte).

We have to start exposing RawMessage to the users, so that we are able to add more information to the message, like metadata.

This commit creates these new interfaces, creates a new engine that handles them (with more robust tests), deprecates old code, and adds an example.